+++ src/main/java/org/owasp/webgoat/lessons/htmltampering/HtmlTampering.java
@@ -1,4 +1,4 @@
-package org.owasp.webgoat.lessons.html_tampering;
+package org.owasp.webgoat.lessons.htmltampering;
 
 import org.owasp.webgoat.container.lessons.Category;
 import org.owasp.webgoat.container.lessons.Lesson;
