+++ src/main/java/org/owasp/webgoat/lessons/htmltampering/HtmlTamperingTask.java
@@ -20,7 +20,7 @@
  * Source for this application is maintained at https://github.com/WebGoat/WebGoat, a repository for free software projects.
  */
 
-package org.owasp.webgoat.lessons.html_tampering;
+package org.owasp.webgoat.lessons.htmltampering;
 
 import org.owasp.webgoat.container.assignments.AssignmentEndpoint;
 import org.owasp.webgoat.container.assignments.AssignmentHints;
