+++ src/main/resources/lessons/htmltampering/i18n/WebGoatLabels.properties
